# More SIMD optimizations (optimize determine_optimal_set_of_endpoint_formats_to_use, compute_ideal_quantized_weights_for_decimation_table)