- Change vecmathlib header to have a separate `vmask` data type that is result of comparison operations, and an argument for `select()`
- Expand vecmathlib with additional operations (gather, horizontal min, extracting lane value etc.)
- Change loop in `determine_optimal_set_of_endpoint_formats_to_use` to be SIMD; this loops 4 times over 2048 sized arrays and was showing up in the profiler for me.
- Change `compute_ideal_quantized_weights_for_decimation_table` hardcoded 4-wide AVX2 loop to use vecmathlib, and so now it uses SIMD also on SSEn.

In my tests, on a 2048x2048 texture with 6x6 block sizes, -medium preset:
- Mac (2018 MacBookPro, Core i9 2.9GHz):
    - SSE2: 6.01s -> 5.20s
    - SSE4.2: 5.61s -> 5.13s
    - AVX2: 5.17s -> 4.74s
- Windows (AMD ThreadRipper 1950X 3.4GHz, 16 threads):
    - SSE2: 3.03s -> 2.89s
    - SSE4.2: 2.74s -> 2.52s
    - AVX2: 2.63s -> 2.59s