@aras-p 2020-10-26
d'oh, sorry! Fixed now.

What are the speedups you see btw? I see the most (as expected) on SSE2/SSE4.2, with older AMD Ryzen being almost no improvement on AVX2 (it's dual-pumped IIRC, so 8-wide is not really faster than 4-wide), but AVX2 on Intel CPU does see a speedup.