@solidpixel 2020-10-26
> What are the speedups you see btw?

Roughly, with AVX2 on an Intel i5-i9600K:

- `-fast`: seeing 1-3% for HDR images and normal maps, 2-7% for LDR image
- `-medium`: seeing 1-2% for HDR images and normal maps, 1-4% for LDR image
- `-thorough`: seeing proportionally less - around 0-2% for LDR images

I've not historically tracked pure SSE2, so I'm expecting great speedups there if we replace all of the existing hand-coded paths with the new vector maths support (given there wasn't an intrinsics path there at all). It won't quite catch SSE 4.2, but it should be pretty close.