@solidpixel 2020-10-26
Here's the full performance report for avx2: [pr166-report.txt](https://github.com/ARM-software/astc-encoder/files/5440955/pr166-report.txt)