@aras-p 2020-10-26
> if we replace all of the existing hand-coded paths

That reminds me -- how many of hardcoded paths there are today, actually? I don't see many -- there's some paths that are writtten in `astcenc_mathlib.h` that are not active (guarded by `&& USE_SCALAR_SSE` which is zero). I only see `flt_rte()` and `flt_rd()` functions that use SSE (4.2).