@solidpixel 2020-10-26
> how many of hardcoded paths there are today, actually?

There weren't many - just three or four actual "algorithms" where we did intrinsics for the whole code path. Grep says that it looks like you might have hit them all already =)

The `USE_SCALAR_SSE` paths were just for testing some algorithmic changes before vectorization - expect the performance sucks for real usage so it's off by default.